refactor: always use native intersection observer
In this commit, we retrieve the native `IntersectionObserver` implementation hidden by
`__zone_symbol__IntersectionObserver`. This would be the unpatched version of the observer
present in zone.js environments.
Otherwise, if the user is using zoneless change detection (and zone.js is not included),
we fall back to the native implementation. Accessing the native implementation allows us
to remove `runOutsideAngular` calls and reduce indentation, making the code a bit more readable.

I also reduced Prettier print width to 90, because 120 seems to be
large (the code overlaps the editor).