feat: allow lazy-loading `tippy.js`
In this commit, we implement lazy loading for `tippy.js` to avoid loading it until it's actually
necessary. Tooltips are commonly used across the app, but they’re not needed until the directive
is rendered. However, the `tippy.js` package is always included in the common bundle.
For example, if we have a `<button [tp]="..." />` element hidden by a conditional
`@if (condition) { ... }`, `tippy.js` would load immediately, even if the button isn't in the DOM.

We’ve added a `loader` property to the config, allowing users to specify a loader function for
the Tippy module. This can be `() => tippy` or `() => import('tippy.js')`, depending on whether
users want to lazy-load it.

This change is breaking because it introduces a required property to the config.
However, the migration should be straightforward, as users only need to explicitly add
the new property to `provideTippyConfig`.