fix: enable strict mode
In this commit, we enable strict mode because it is required to generate correct type definition
(`.d.ts`) files. These files differ depending on whether strict mode is enabled or disabled.
Enabling strict mode allows `null|undefined` values to be provided by the parent component
(e.g., `[tp]="null"`, which should be allowed), but the compiler removes `null|undefined`
values from the binding.

We also had to patch the `@popperjs/core` types due to a mistake in their definitions.
This patch was necessary for successful compilation