feat: add `@ngneat/helipopper/config`
In this commit, we add a new subpackage called `@ngneat/helipopper/config`.
This change is intended to separate configuration logic from UI-related functionality.
The configuration should be imported at the root level, while the UI components should
only be imported when necessary. Previously, importing the library at the root
level (when providing the configuration) caused the bundler to include everything in the
main bundle immediately.

With this approach, the configuration is bundled separately at the root level and
includes only the provider functions.

Additionally, while we had already added a `loader` function, to avoid breaking the
`provideTippyConfig` signature, we introduced a separate function called `provideTippyLoader`,
which exclusively accepts a loader function.

The `provideTippyConfig` has been reverted to return a provider (not an environment provider),
becuase it might be provided at the component level too.